Solve package naming conflict Add BinaryReader Add Test on binary reader
