+++ simulators/cycle_accurate_simulator/src/test/scala/simulatorTest/gemm/TensorGemmTest.scala
@@ -1,4 +1,4 @@
-package simulator.gemm
+package simulatorTest.gemm
 
 import chisel3._
 import chisel3.util._
